Combine both SFMLtreeDataProvider and MixedTreeDataProvider in one class for the activity bar views, looks and works identical as before the fusion
